+++ lib/hashdiff_sym/util.rb
@@ -1,4 +1,4 @@
-module HashDiff
+module HashDiffSym
 
   # @private
   #