+++ spec/hashdiff_sym/diff_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe HashDiff do
+describe HashDiffSym do
   it "should be able to diff two empty hashes" do
-    diff = HashDiff.diff({}, {})
+    diff = HashDiffSym.diff({}, {})
     diff.should == []
   end
 
@@ -10,41 +10,41 @@ describe HashDiff do
     a = { 'a' => 3, 'b' => 2 }
     b = {}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [['-', 'a', 3], ['-', 'b', 2]]
 
-    diff = HashDiff.diff(b, a)
+    diff = HashDiffSym.diff(b, a)
     diff.should == [['+', 'a', 3], ['+', 'b', 2]]
   end
 
   it "should be able to diff two equal hashes" do
-    diff = HashDiff.diff({ 'a' => 2, 'b' => 2}, { 'a' => 2, 'b' => 2 })
+    diff = HashDiffSym.diff({ 'a' => 2, 'b' => 2}, { 'a' => 2, 'b' => 2 })
     diff.should == []
   end
 
   it "should be able to diff two hashes with equivalent numerics, when strict is false" do
-    diff = HashDiff.diff({ 'a' => 2.0, 'b' => 2 }, { 'a' => 2, 'b' => 2.0 }, :strict => false)
+    diff = HashDiffSym.diff({ 'a' => 2.0, 'b' => 2 }, { 'a' => 2, 'b' => 2.0 }, :strict => false)
     diff.should == []
   end
 
   it "should be able to diff changes in hash value" do
-    diff = HashDiff.diff({ 'a' => 2, 'b' => 3, 'c' => " hello" }, { 'a' => 2, 'b' => 4, 'c' => "hello" })
+    diff = HashDiffSym.diff({ 'a' => 2, 'b' => 3, 'c' => " hello" }, { 'a' => 2, 'b' => 4, 'c' => "hello" })
     diff.should == [['~', 'b', 3, 4], ['~', 'c', " hello", "hello"]]
   end
 
   it "should be able to diff changes in hash value which is array" do
-    diff = HashDiff.diff({ 'a' => 2, 'b' => [1, 2, 3] }, { 'a' => 2, 'b' => [1, 3, 4]})
+    diff = HashDiffSym.diff({ 'a' => 2, 'b' => [1, 2, 3] }, { 'a' => 2, 'b' => [1, 3, 4]})
     diff.should == [['-', 'b[1]', 2], ['+', 'b[2]', 4]]
   end
 
   it "should be able to diff changes in hash value which is hash" do
-    diff = HashDiff.diff({ 'a' => { 'x' => 2, 'y' => 3, 'z' => 4 }, 'b' => { 'x' => 3, 'z' => 45 } },
+    diff = HashDiffSym.diff({ 'a' => { 'x' => 2, 'y' => 3, 'z' => 4 }, 'b' => { 'x' => 3, 'z' => 45 } },
                          { 'a' => { 'y' => 3 }, 'b' => { 'y' => 3, 'z' => 30 } })
     diff.should == [['-', 'a.x', 2], ['-', 'a.z', 4], ['-', 'b.x', 3], ['~', 'b.z', 45, 30], ['+', 'b.y', 3]]
   end
 
   it "should be able to diff similar objects in array" do
-    diff = HashDiff.best_diff({ 'a' => [{ 'x' => 2, 'y' => 3, 'z' => 4 }, { 'x' => 11, 'y' => 22, 'z' => 33 }], 'b' => { 'x' => 3, 'z' => 45 } },
+    diff = HashDiffSym.best_diff({ 'a' => [{ 'x' => 2, 'y' => 3, 'z' => 4 }, { 'x' => 11, 'y' => 22, 'z' => 33 }], 'b' => { 'x' => 3, 'z' => 45 } },
                               { 'a' => [{ 'y' => 3 }, { 'x' => 11, 'z' => 33 }], 'b' => { 'y' => 22 } })
     diff.should == [['-', 'a[0].x', 2], ['-', 'a[0].z', 4], ['-', 'a[1].y', 22], ['-', 'b.x', 3], ['-', 'b.z', 45], ['+', 'b.y', 22]]
   end
@@ -53,10 +53,10 @@ describe HashDiff do
     a = {"a"=>3, "c"=>11, "d"=>45, "e"=>100, "f"=>200}
     b = {"a"=>3, "c"=>11, "d"=>45, "e"=>100, "f"=>200, "g"=>300}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [['+', 'g', 300]]
 
-    diff = HashDiff.diff(b, a)
+    diff = HashDiffSym.diff(b, a)
     diff.should == [['-', 'g', 300]]
   end
 
@@ -64,10 +64,10 @@ describe HashDiff do
     a = {"a" => 3}
     b = {"a" => {"a1" => 1, "a2" => 2}}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should  == [['~', 'a', 3, {"a1" => 1, "a2" => 2}]]
 
-    diff = HashDiff.diff(b, a)
+    diff = HashDiffSym.diff(b, a)
     diff.should  == [['~', 'a', {"a1" => 1, "a2" => 2}, 3]]
   end
 
@@ -75,7 +75,7 @@ describe HashDiff do
     a = {"a" => 1, "b" => [1, 2]}
     b = {"a" => 1, "b" => []}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [['-', 'b[1]', 2], ['-', 'b[0]', 1]]
   end
 
@@ -83,7 +83,7 @@ describe HashDiff do
     a = {"a" => 1, "b" => [1, 2]}
     b = {"a" => 1, "b" => nil}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [["~", "b", [1, 2], nil]]
   end
 
@@ -91,7 +91,7 @@ describe HashDiff do
     a = {"a" => 1, "b" => [1, 2]}
     b = {"a" => 1}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [["-", "b", [1, 2]]]
   end
 
@@ -99,7 +99,7 @@ describe HashDiff do
     a = {"a" => 1, "b" => {"b1" => 1, "b2" =>2}}
     b = {"a" => 1}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [["-", "b", {"b1"=>1, "b2"=>2}]]
   end
 
@@ -107,7 +107,7 @@ describe HashDiff do
     a = {"a" => 1, "b" => {"b1" => 1, "b2" =>2}}
     b = {"a" => 1, "b" => {}}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [['-', 'b.b1', 1], ['-', 'b.b2', 2]]
   end
 
@@ -115,7 +115,7 @@ describe HashDiff do
     a = {"a" => 1, "b" => {"b1" => 1, "b2" =>2}}
     b = {"a" => 1, "b" => nil}
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [["~", "b", {"b1"=>1, "b2"=>2}, nil]]
   end
 
@@ -123,7 +123,7 @@ describe HashDiff do
     a = [{'a' => 1, 'b' => 2, 'c' => 3, 'd' => 4, 'e' => 5}, 3]
     b = [1, {'a' => 1, 'b' => 2, 'c' => 3, 'e' => 5}]
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [['-', '[0].d', 4], ['+', '[0]', 1], ['-', '[2]', 3]]
   end
 
@@ -131,7 +131,7 @@ describe HashDiff do
     a = [{'a' => 1, 'b' => 2, 'c' => 3, 'd' => 4, 'e' => 5}, {'x' => 5, 'y' => 6, 'z' => 3}, 3]
     b = [{'a' => 1, 'b' => 2, 'c' => 3, 'e' => 5}, 3]
 
-    diff = HashDiff.diff(a, b)
+    diff = HashDiffSym.diff(a, b)
     diff.should == [["-", "[0].d", 4], ["-", "[1]", {"x"=>5, "y"=>6, "z"=>3}]]
   end
 
@@ -139,12 +139,12 @@ describe HashDiff do
     a = [{'a' => 1, 'b' => 2, 'c' => 3, 'd' => 4, 'e' => 5}, {'x' => 5, 'y' => 6, 'z' => 3}, 3]
     b = [{'a' => 1, 'b' => 2, 'c' => 3, 'e' => 5}, 3]
 
-    diff = HashDiff.diff(a, b, :similarity => 0.8, :delimiter => "\t")
+    diff = HashDiffSym.diff(a, b, :similarity => 0.8, :delimiter => "\t")
     diff.should == [["-", "[0]\td", 4], ["-", "[1]", {"x"=>5, "y"=>6, "z"=>3}]]
   end
 
   it "should transform symbols to string correctly" do
-    diff = HashDiff.diff({ 'a' => { x: 2, y: 3, z: 4 }, 'b' => { x: 3, z: [1, 2, 3] } },
+    diff = HashDiffSym.diff({ 'a' => { x: 2, y: 3, z: 4 }, 'b' => { x: 3, z: [1, 2, 3] } },
                          { 'a' => { y: 3 }, 'b' => { y: 3, z: [2, 3, 4] } })
     diff.should == [['-', 'a.:x', 2], ['-', 'a.:z', 4], ['-', 'b.:x', 3], ["-", "b.:z[0]", 1], ["+", "b.:z[2]", 4], ['+', 'b.:y', 3]]
   end
@@ -154,10 +154,10 @@ describe HashDiff do
       a = {'a' => 0.558, 'b' => 0.0, 'c' => 0.65, 'd' => 'fin'}
       b = {'a' => 0.557, 'b' => 'hats', 'c' => 0.67, 'd' => 'fin'}
 
-      diff = HashDiff.diff(a, b, :numeric_tolerance => 0.01)
+      diff = HashDiffSym.diff(a, b, :numeric_tolerance => 0.01)
       diff.should == [["~", "b", 0.0, 'hats'], ["~", "c", 0.65, 0.67]]
 
-      diff = HashDiff.diff(b, a, :numeric_tolerance => 0.01)
+      diff = HashDiffSym.diff(b, a, :numeric_tolerance => 0.01)
       diff.should == [["~", "b", 'hats', 0.0], ["~", "c", 0.67, 0.65]]
     end
 
@@ -165,10 +165,10 @@ describe HashDiff do
       a = {'a' => {'x' => 0.4, 'y' => 0.338}, 'b' => [13, 68.03]}
       b = {'a' => {'x' => 0.6, 'y' => 0.341}, 'b' => [14, 68.025]}
 
-      diff = HashDiff.diff(a, b, :numeric_tolerance => 0.01)
+      diff = HashDiffSym.diff(a, b, :numeric_tolerance => 0.01)
       diff.should == [["~", "a.x", 0.4, 0.6], ["-", "b[0]", 13], ["+", "b[0]", 14]]
 
-      diff = HashDiff.diff(b, a, :numeric_tolerance => 0.01)
+      diff = HashDiffSym.diff(b, a, :numeric_tolerance => 0.01)
       diff.should == [["~", "a.x", 0.6, 0.4], ["-", "b[0]", 14], ["+", "b[0]", 13]]
     end
   end
@@ -177,14 +177,14 @@ describe HashDiff do
     it "should strip strings before comparing" do
       a = { 'a' => " foo", 'b' => "fizz buzz"}
       b = { 'a' => "foo", 'b' => "fizzbuzz"}
-      diff = HashDiff.diff(a, b, :strip => true)
+      diff = HashDiffSym.diff(a, b, :strip => true)
       diff.should == [['~', 'b', "fizz buzz", "fizzbuzz"]]
     end
 
     it "should strip nested strings before comparing" do
       a = { 'a' => { 'x' => " foo" }, 'b' => ["fizz buzz", "nerf"] }
       b = { 'a' => { 'x' => "foo" }, 'b' => ["fizzbuzz", "nerf"] }
-      diff = HashDiff.diff(a, b, :strip => true)
+      diff = HashDiffSym.diff(a, b, :strip => true)
       diff.should == [['-', 'b[0]', "fizz buzz"], ['+', 'b[0]', "fizzbuzz"]]
     end
   end
@@ -193,14 +193,14 @@ describe HashDiff do
     it "should strip strings before comparing" do
       a = { 'a' => "Foo", 'b' => "fizz buzz"}
       b = { 'a' => "foo", 'b' => "fizzBuzz"}
-      diff = HashDiff.diff(a, b, :case_insensitive => true)
+      diff = HashDiffSym.diff(a, b, :case_insensitive => true)
       diff.should == [['~', 'b', "fizz buzz", "fizzBuzz"]]
     end
 
     it "should ignore case on nested strings before comparing" do
       a = { 'a' => { 'x' => "Foo" }, 'b' => ["fizz buzz", "nerf"] }
       b = { 'a' => { 'x' => "foo" }, 'b' => ["fizzbuzz", "nerf"] }
-      diff = HashDiff.diff(a, b, :case_insensitive => true)
+      diff = HashDiffSym.diff(a, b, :case_insensitive => true)
       diff.should == [['-', 'b[0]', "fizz buzz"], ['+', 'b[0]', "fizzbuzz"]]
     end
   end
@@ -209,7 +209,7 @@ describe HashDiff do
     it 'should apply filters to proper object types' do
       a = { 'a' => " foo", 'b' => 35, 'c' => 'bar', 'd' => 'baz' }
       b = { 'a' => "foo", 'b' => 35.005, 'c' => 'bar', 'd' => 18.5}
-      diff = HashDiff.diff(a, b, :strict => false, :numeric_tolerance => 0.01, :strip => true)
+      diff = HashDiffSym.diff(a, b, :strict => false, :numeric_tolerance => 0.01, :strip => true)
       diff.should == [['~', 'd', "baz", 18.5]]
     end
   end
@@ -218,7 +218,7 @@ describe HashDiff do
     it "should apply both filters to strings" do
       a = { 'a' => " Foo", 'b' => "fizz buzz"}
       b = { 'a' => "foo", 'b' => "fizzBuzz"}
-      diff = HashDiff.diff(a, b, :case_insensitive => true, :strip => true)
+      diff = HashDiffSym.diff(a, b, :case_insensitive => true, :strip => true)
       diff.should == [['~', 'b', "fizz buzz", "fizzBuzz"]]
     end
   end
@@ -228,7 +228,7 @@ describe HashDiff do
     let(:b) { { 'a' => 'bus', 'b' => 'truck', 'c' => ' plan'} }
 
     it 'should compare using proc specified in block' do
-      diff = HashDiff.diff(a, b) do |prefix, obj1, obj2|
+      diff = HashDiffSym.diff(a, b) do |prefix, obj1, obj2|
         case prefix
         when /a|b|c/
           obj1.length == obj2.length
@@ -241,7 +241,7 @@ describe HashDiff do
       x = { 'a' => 'car', 'b' => 'boat'}
       y = { 'a' => 'car' }
 
-      diff = HashDiff.diff(x, y) do |prefix, obj1, obj2|
+      diff = HashDiffSym.diff(x, y) do |prefix, obj1, obj2|
         case prefix
         when /b/
           true
@@ -251,7 +251,7 @@ describe HashDiff do
     end
 
     it 'should compare with both proc and :strip when both provided' do
-      diff = HashDiff.diff(a, b, :strip => true) do |prefix, obj1, obj2|
+      diff = HashDiffSym.diff(a, b, :strip => true) do |prefix, obj1, obj2|
         case prefix
         when 'a'
           obj1.length == obj2.length